+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.2 - in progress
 - Start applying Bootstrap styles.
+- Add a minimalist browse page.
 
 ## [0.0.1](https://github.com/hubmapconsortium/flask-data-portal/tree/v0.0.1) - 2019-10-11
 ### Added